requirements.txt
#download model 
#gdown --id 1xYuk8x1rh8xhUYuDYAq-HYp3Jt0DaARG 
#gdown --id 1eaDJj5DOBQEjiCX9a9uWBijSGS15sXkk 
# install cuddnn 
#cp -P lib/* /usr/local/cuda/lib64/ 
#cp -P include/* /usr/local/cuda/include/ 
#chmod a+r /usr/local/cuda/include/cudnn*.h /usr/local/cuda/lib64/libcudnn*
# install dlib dependence 
# apt-get update 
# apt-get install build-essential cmake -y 
# apt-get install libopenblas-dev liblapack-dev -y 
# apt-get install libx11-dev libgtk-3-dev -y 
# apt-get install python3 python3-dev python3-pip -y 
flask 
flask-cors 
flask-socketio
imutils 
dlib 
Keras==2.13.1 
tensorflow==2.13.0 
scipy 
Keras-Applications==1.0.8 
Keras-Preprocessing==1.1.2 
Pillow==8.2.0 
mtcnn 
torch
shapely==2.0.6 
pyclipper 
scikit-image 
imgaug 
lmdb 
tqdm 
loguru 
matplotlib 
paddlepaddle-gpu
lanms-nova
#ngrok
#curl -sSL https://ngrok-agent.s3.amazonaws.com/ngrok.asc \
# 	| tee /etc/apt/trusted.gpg.d/ngrok.asc >/dev/null \
# 	&& echo "deb https://ngrok-agent.s3.amazonaws.com buster main" \
# 	| tee /etc/apt/sources.list.d/ngrok.list \
# 	&& apt update \
# 	&& apt install ngrok

# ngrok config add-authtoken 2njSZOlbuhRX3Kcb3Yiyl7ZBuKq_7K7gzmJYwyGUmua4kiCHc
# ngrok http http://localhost:5000